Refactor integration test names and packages
